+++ backend/controllers/user.js
@@ -1,9 +1,24 @@
 const { StatusCodes } = require("http-status-codes");
 const CustomError = require("../errors");
 const User = require("../models/user");
+const {
+  UnauthenticatedError,
+  NotFoundError,
+
+  UnauthorizedError,
+} = require("../errors");
 const getAllUsers = async (req, res) => {
+  if (!req.user) {
+    throw new UnauthenticatedError("Your not authenticated");
+  }
+  if (!req.user.isAdmin) {
+    throw new UnauthorizedError("Route protected to admins only");
+  }
   const users = await User.find({}).select("-password"); // Exclude password field
 
+  if (user.length === 0) {
+    throw new NotFoundError("No users found");
+  }
   res.status(200).json({
     users: users.map((user) => ({
       id: user._id,
@@ -16,6 +31,10 @@ const getAllUsers = async (req, res) => {
 };
 
 const getUser = async (req, res) => {
+  if (!req.user) {
+    throw new UnauthenticatedError("Your not authenticated");
+  }
+
   const id = req.params.id;
   const user = await User.findById(id);
   if (!user) {
@@ -30,6 +49,12 @@ const getUser = async (req, res) => {
   });
 };
 const deleteUser = async (req, res) => {
+  if (!req.user) {
+    throw new UnauthenticatedError("Your not authenticated");
+  }
+  if (!req.user.isAdmin) {
+    throw new UnauthorizedError("Route protected to admins only");
+  }
   const id = req.params.id;
   const user = await User.findOneAndDelete({ _id: id });
   if (!user) {
@@ -39,4 +64,25 @@ const deleteUser = async (req, res) => {
     .status(StatusCodes.OK)
     .json({ message: `user ${id} delete successfully` });
 };
-module.exports = { getAllUsers, getUser, deleteUser };
+const updateUser = async (req, res) => {
+  if (!req.user) {
+    throw new UnauthenticatedError("Your not authenticated");
+  }
+  if (!req.user.isAdmin) {
+    throw new UnauthorizedError("Route protected to admins only");
+  }
+  const { id: userID } = req.params;
+
+  const user = await User.findOneAndUpdate({ _id: userID }, req.body, {
+    new: true,
+    runValidators: true,
+  });
+
+  if (!user) {
+    throw new NotFoundError(`No user with id : ${userID}`);
+  }
+
+  res.status(200).json({ user });
+};
+
+module.exports = { getAllUsers, getUser, deleteUser, updateUser };
